\name{bmixture-internal}
\alias{bmixture-internal}

\alias{bmixnorm_unknown_k}
\alias{bmixnorm_fixed_k}
\alias{bmixgamma_unknown_k}
\alias{bmixgamma_fixed_k}

\docType{data}

\title{ Internal bmixture functions and datasets }

\description{ Nothing to see here move along }

\author{ Abdolreza Mohammadi <a.mohammadi@uvt.nl> }

\keyword{ internal }
