\name{bmixture-package}
\alias{bmixture-package}
\docType{package}

\title{ Bayesian Estimation for Finite Mixture of Distributions }

\description{
The \code{R} package \pkg{bmixture} provides statistical tools for Bayesian estimation in finite mixture of distributions.       
The package implemented the improvements in the Bayesian literature, including Mohammadi and Salehi-Rad (2012) and Mohammadi et al. (2013).
Besides, the package contains several functions for simulation and visualization, as well as a real dataset taken from the literature.
}

\author{Abdolreza Mohammadi <a.mohammadi@uvt.nl>}

\references{
Mohammadi, A., M. R. Salehi-Rad, and E. C. Wit (2013) Using mixture of Gamma distributions for Bayesian analysis in an M/G/1 queue with optional second service. \emph{Computational Statistics} 28(2):683-700

Mohammadi, A, and M. R. Salehi-Rad (2012) Bayesian inference and prediction in an M/G/1 with optional second service. \emph{Communications in Statistics-Simulation and Computation} 41(3):419-435

Stephens, M. (2000) Bayesian analysis of mixture models with an unknown number of components-an alternative to reversible jump methods. \emph{Annals of statistics} 28(1):40-74
}

\keyword{ package }
